ctrip-assistant-mcp/pyproject.toml
[project]
name = "ctrip-assistant-mcp"
version = "0.1.0"
description = "携程商旅智能体（航班/酒店预订）"
authors = [{ name = "Your Name", email = "your@email.com" }]
requires-python = ">=3.13"
dependencies = [
  # LangChain & LangGraph 1.0+
  "langchain>=1.0.0",
  "langchain-core>=1.0.0",
  "langgraph>=1.0.0",
  "langgraph-checkpoint-sqlite>=1.0.0",
  "langchain-tavily>=0.0.1",
  # FastAPI & 服务部署
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  # SQLAlchemy
  "sqlalchemy>=2.0.0",
  # 配置解析
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  # 其他基础依赖
  "pydantic>=2.0.0",
  "requests>=2.31.0",
  "dashscope>=1.20.0",
  "pandas>=2.0.0",
  "gradio>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app","config"]
